The error occurs in this line of code:
ExamEntity examSelected = exams.get(i);

This is because the index being accessed (i) is out of bounds of the exams list.

To fix this, you should check if the index is within the bounds of the list before attempting to access it. You can do this with the following code:

if (i >= 0 && i < exams.size()) {
    ExamEntity examSelected = exams.get(i);
    // rest of code
}